Enable packaging of the collector
The packaged version of this code should be trackable like any
other in the Orcfax stack. We re-implement packaging here.
