+++ .gitignore
@@ -132,3 +132,12 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+# Cardano
+*.addr
+*.vkey
+*.skey
+
+# Custom
+*.bk
+cnt/*
